refactor: server - extract Message model
