+++ chat/src/main/java/star/astro/chat/model/ChatMessage.java
@@ -1,14 +1,10 @@
 package star.astro.chat.model;
 
-public class ChatMessage {
+public class ChatMessage extends Message {
 
-    private String sender;
-    private String receiver;
     private String content;
-    private String time;
 
     public ChatMessage() {
-
     }
 
     public ChatMessage(String sender, String receiver, String content, String time) {
@@ -18,36 +14,12 @@ public class ChatMessage {
         this.time = time;
     }
 
-    public String getSender() {
-        return sender;
-    }
-
-    public void setSender(String sender) {
-        this.sender = sender;
-    }
-
-    public String getReceiver() {
-        return receiver;
-    }
-
-    public void setReceiver(String receiver) {
-        this.receiver = receiver;
-    }
-
     public String getContent() {
         return content;
     }
 
     public void setContent(String content) {
         this.content = content;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public void setTime(String time) {
-        this.time = time;
     }
 
 }
