+++ chat/src/main/java/star/astro/chat/model/Notification.java
@@ -1,11 +1,8 @@
 package star.astro.chat.model;
 
-public class Notification {
+public class Notification extends Message {
 
-    private String sender;
-    private String receiver;
     private int type;
-    private String time;
 
     public Notification() {
 
@@ -18,36 +15,12 @@ public class Notification {
         this.time = time;
     }
 
-    public String getSender() {
-        return sender;
-    }
-
-    public void setSender(String sender) {
-        this.sender = sender;
-    }
-
-    public String getReceiver() {
-        return receiver;
-    }
-
-    public void setReceiver(String receiver) {
-        this.receiver = receiver;
-    }
-
     public int getType() {
         return type;
     }
 
     public void setType(int type) {
         this.type = type;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public void setTime(String time) {
-        this.time = time;
     }
 
 }
